Fix some qreal issues.
CCBUG: 342642

floating point values which count pixels or screen locations
can and should be qreal.

floating point values which are
involved with computational manipulation of the data
should be double, even though this is slower in ARM.

They shouldn't mix if it can be helped.

+++ src/libkst/editablevector.cpp
@@ -59,7 +59,7 @@ void EditableVector::setValue(const int &i, const double &val) { //sa Vector::ch
     _scalars["sumsquared"]->setValue(_sum*_sum);
     _scalars["max"]->setValue(qMax(_max,val));
     _scalars["min"]->setValue(qMin(_min,_min));
-    double b=(float)(qMax((float)0.0f,(float)_minPos));
+    double b=qMax(double(0.0),_minPos);
     _scalars["minpos"]->setValue(qMin(_min,b));
     _scalars["last"]->setValue(_v[_size-1]);
     _scalars["first"]->setValue(_v[0]);
